Updating name of API endpoint. Creating String extension methods
